# studio/openapi.json
{
  "components": {
    "schemas": {
      "ApiErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ApiErrorIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ApiErrorDetail",
        "type": "object"
      },
      "ApiErrorIssue": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "location": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "location",
          "code",
          "message"
        ],
        "title": "ApiErrorIssue",
        "type": "object"
      },
      "ApiErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "CreateProjectRequest": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "minLength": 1,
            "title": "Actor",
            "type": "string"
          },
          "collaboration_mode": {
            "default": "managed",
            "enum": [
              "managed",
              "co_design",
              "governance"
            ],
            "title": "Collaboration Mode",
            "type": "string"
          },
          "created_from": {
            "items": {
              "$ref": "#/components/schemas/SourceReference"
            },
            "title": "Created From",
            "type": "array"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "name",
          "actor",
          "reason"
        ],
        "title": "CreateProjectRequest",
        "type": "object"
      },
      "CreateProposalJobRequest": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "minLength": 1,
            "title": "Actor",
            "type": "string"
          },
          "kind": {
            "enum": [
              "product_intent",
              "problem_model",
              "outcome_contract",
              "product_theses",
              "web_generation_contract"
            ],
            "title": "Kind",
            "type": "string"
          },
          "raw_input": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Input"
          },
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "actor",
          "reason"
        ],
        "title": "CreateProposalJobRequest",
        "type": "object"
      },
...
      "ProductProjectResponse": {
        "additionalProperties": false,
        "properties": {
          "collaboration_mode": {
            "enum": [
              "managed",
              "co_design",
              "governance"
            ],
            "title": "Collaboration Mode",
            "type": "string"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "created_from": {
            "items": {
              "$ref": "#/components/schemas/SourceReference"
            },
            "title": "Created From",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/RevisionMetaResponse"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "paused",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "revision_id",
          "meta",
          "name",
          "collaboration_mode",
          "status",
          "created_from",
          "content_hash"
        ],
        "title": "ProductProjectResponse",
        "type": "object"
      },
...
      "ProductProposalJobResponse": {
        "additionalProperties": false,
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          },
          "fingerprint": {
            "title": "Fingerprint",
            "type": "string"
          },
          "input_dependencies": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Input Dependencies",
            "type": "array"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "product_intent",
              "problem_model",
              "outcome_contract",
              "product_theses",
              "web_generation_contract"
            ],
            "title": "Kind",
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/RevisionMetaResponse"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "provider": {
            "const": "deterministic_fake",
            "title": "Provider",
            "type": "string"
          },
          "provider_version": {
            "title": "Provider Version",
            "type": "string"
          },
          "raw_input": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Input"
          },
          "result_expected_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Expected Revision"
          },
          "result_object_id": {
            "title": "Result Object Id",
            "type": "string"
          },
          "result_object_ids": {
            "items": {
              "type": "string"
            },
            "title": "Result Object Ids",
            "type": "array"
          },
          "result_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Revision Id"
          },
          "result_revision_ids": {
            "items": {
              "type": "string"
            },
            "title": "Result Revision Ids",
            "type": "array"
          },
          "revision_id": {
            "title": "Revision Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "stale_input",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "revision_id",
          "meta",
          "project_id",
          "kind",
          "status",
          "provider",
          "provider_version",
          "input_dependencies",
          "result_object_id",
          "result_object_ids",
          "result_expected_revision",
          "raw_input",
          "fingerprint",
          "attempt",
          "result_revision_id",
          "result_revision_ids",
          "error_code",
          "error_summary",
          "content_hash"
        ],
        "title": "ProductProposalJobResponse",
        "type": "object"
      },
...
      "ProposalJobActionRequest": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "minLength": 1,
            "title": "Actor",
            "type": "string"
          }
        },
        "required": [
          "actor"
        ],
        "title": "ProposalJobActionRequest",
        "type": "object"
      },
...
      "RevisionMetaResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "parent_revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Revision Id"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "revision",
          "parent_revision_id",
          "created_at",
          "created_by",
          "reason"
        ],
        "title": "RevisionMetaResponse",
        "type": "object"
      },
      "SourceReference": {
        "additionalProperties": false,
        "description": "A compact pointer to the origin of a product-domain statement.",
        "properties": {
          "locator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locator"
          },
          "revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Id"
          },
          "source_id": {
            "minLength": 1,
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "enum": [
              "user_input",
              "imported_artifact",
              "research",
              "evidence",
              "tool_result",
              "model_proposal",
              "human_decision"
            ],
            "title": "Source Type",
            "type": "string"
          }
        },
        "required": [
          "source_type",
          "source_id"
        ],
        "title": "SourceReference",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first command and projection API for Product Studio. Proposal and generation work is created as a persisted job; an explicit run command performs only the bounded local step. Generated source is not installed, executed, or exposed as a preview.",
    "title": "psyteardown Product Studio API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects": {
      "post": {
        "operationId": "create_project_api_v1_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Revision or domain-state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request or domain validation failed"
          }
        },
        "summary": "Create Project",
        "tags": [
          "product-projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/proposal-jobs": {
      "get": {
        "operationId": "list_proposal_jobs_api_v1_projects__project_id__proposal_jobs_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductProposalJobResponse"
                  },
                  "title": "Response List Proposal Jobs Api V1 Projects  Project Id  Proposal Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Revision or domain-state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request or domain validation failed"
          }
        },
        "summary": "List Proposal Jobs",
        "tags": [
          "product-proposal-jobs"
        ]
      },
      "post": {
        "operationId": "create_proposal_job_api_v1_projects__project_id__proposal_jobs_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProposalJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductProposalJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Revision or domain-state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request or domain validation failed"
          }
        },
        "summary": "Create Proposal Job",
        "tags": [
          "product-proposal-jobs"
        ]
      }
    },
    "/api/v1/projects/{project_id}/proposal-jobs/{job_id}": {
      "get": {
        "operationId": "get_proposal_job_api_v1_projects__project_id__proposal_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductProposalJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Revision or domain-state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request or domain validation failed"
          }
        },
        "summary": "Get Proposal Job",
        "tags": [
          "product-proposal-jobs"
        ]
      }
    },
    "/api/v1/projects/{project_id}/proposal-jobs/{job_id}/retries": {
      "post": {
        "operationId": "retry_proposal_job_api_v1_projects__project_id__proposal_jobs__job_id__retries_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalJobActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductProposalJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Revision or domain-state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request or domain validation failed"
          }
        },
        "summary": "Retry Proposal Job",
        "tags": [
          "product-proposal-jobs"
        ]
      }
    },
    "/api/v1/projects/{project_id}/proposal-jobs/{job_id}/runs": {
      "post": {
        "operationId": "run_proposal_job_api_v1_projects__project_id__proposal_jobs__job_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalJobActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductProposalJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Revision or domain-state conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request or domain validation failed"
          }
        },
        "summary": "Run Proposal Job",
        "tags": [
          "product-proposal-jobs"
        ]
      }
    },
...
  }
}